feat(router): able to provide base URL with APP_BASE_HREF
Allows HashLocationStrategy to generate external URL with given
APP_BASE_HREF. This PR should resolve the deploying host page and rest
static files separately issue mentioned in #13482 and #20033.

Modified HashLocationStrategy will take base URL with the following order:

* APP_BASE_HREF
* path of current URL

The HashLocationStrategy used to push only hash part (eg: #/ui/path) to
history API. This will make browser change the host page part of URL in
address bar to path pointed by base-href instead of path of host page.

With this patch, the HashLocationStrategy will push both path and hash
part to history API. In this way the browser will be able to keep the host
page part of URL in address bar to the path of host page.

This is a breaking change. If APP_BASE_HREF is not given, pathname will
be use.